refactor(mcp): carve workspace, config IO, and file IO out of server_support
WorkspaceLayout with its path jail and the derived workflow maps move to
workspace.py; YAML validate/lint/write move to config_io.py; bounded
text readers and suffix helpers move to file_io.py -- all verbatim.
server_support.py re-exports every moved name so existing imports and
tests keep resolving from the old home.
