+++ konfai-mcp/konfai_mcp/server_support.py
@@ -20,24 +20,28 @@ import ast
 import importlib
 import inspect
 import math
-import os
 import re
 import shutil
 from collections.abc import Iterator, Sequence
-from dataclasses import dataclass
-from io import StringIO
 from pathlib import Path
 from typing import Any
 
 import numpy as np
-from ruamel.yaml import YAML
 
-from konfai_mcp.workflows import WORKFLOW_SPECS
-
-YAML_SAFE = YAML(typ="safe")
-YAML_DUMP = YAML()
-YAML_DUMP.default_flow_style = False
-YAML_DUMP.width = 4096
+from .config_io import YAML_DUMP as YAML_DUMP
+from .config_io import YAML_SAFE
+from .config_io import _lint_config_data as _lint_config_data
+from .config_io import validate_yaml_content as validate_yaml_content
+from .config_io import write_config as write_config
+from .config_io import yaml_dump_content as yaml_dump_content
+from .file_io import basename_without_suffixes as basename_without_suffixes
+from .file_io import full_suffix as full_suffix
+from .file_io import read_dataset_sidecar as read_dataset_sidecar
+from .file_io import read_text as read_text
+from .file_io import read_text_range as read_text_range
+from .file_io import read_text_tail as read_text_tail
+from .workspace import WORKFLOW_CONFIG_FILES, WORKFLOW_ROOT_KEYS
+from .workspace import WorkspaceLayout as WorkspaceLayout
 
 # --- Payload float economy -------------------------------------------------
 # Round display floats (metrics, statistics) to a fixed number of significant figures
@@ -139,10 +143,6 @@ def aggregate_case_statistics(stats: dict[str, dict[str, Any]]) -> dict[str, dic
     return result
 
 
-WORKFLOW_CONFIG_FILES = {kind: spec.config_file for kind, spec in WORKFLOW_SPECS.items()}
-
-WORKFLOW_ROOT_KEYS = {kind: spec.root_key for kind, spec in WORKFLOW_SPECS.items()}
-
 CONFIG_AUTHORING_QUESTIONS = [
     "Ask only if uncertain: which dataset groups are inputs, targets, or support-only groups for this task?",
     "Ask only if uncertain: which workflows are intended now: train, prediction, and/or evaluation?",
@@ -153,416 +153,6 @@ CONFIG_AUTHORING_QUESTIONS = [
         "or should one split/root be used for train, test, or evaluation?"
     ),
 ]
-
-
-@dataclass(frozen=True)
-class WorkspaceLayout:
-    """Filesystem layout helper for KonfAI MCP datasets, sessions, and one mutable session workspace."""
-
-    root: Path
-    current_session: str | None = None
-
-    def __post_init__(self) -> None:
-        root = self.root.expanduser().resolve()
-        object.__setattr__(self, "root", root)
-        object.__setattr__(self, "current_session", self._resolve_current_session(self.current_session))
-
-    def sanitize_name(self, name: str) -> str:
-        safe = re.sub(r"[^a-zA-Z0-9._-]+", "_", name.strip())
-        if safe in {"", ".", ".."}:
-            raise ValueError("Name is empty or invalid after sanitization.")
-        return safe
-
-    def _resolve_current_session(self, value: str | None) -> str:
-        if value is not None:
-            return self.sanitize_name(value)
-        env_value = os.environ.get("KONFAI_MCP_SESSION")
-        if env_value:
-            return self.sanitize_name(env_value)
-        marker_path = self.session_marker_path()
-        if marker_path.exists():
-            marker_value = marker_path.read_text(encoding="utf-8").strip()
-            if marker_value:
-                return self.sanitize_name(marker_value)
-        return "default"
-
-    def internal_root_dir(self) -> Path:
-        return self.root / ".konfai_mcp"
-
-    def session_marker_path(self) -> Path:
-        return self.internal_root_dir() / "current_session.txt"
-
-    def apps_catalog_path(self) -> Path:
-        """Editable per-root catalogue of app sources (``{"apps": [...]}``) shared across sessions."""
-        return self.root / "apps_catalog.json"
-
-    def sessions_root(self) -> Path:
-        return self.root / "sessions"
-
-    def session_dir(self, name: str | None = None) -> Path:
-        session_name = self.sanitize_name(name or self.current_session or "default")
-        return self.sessions_root() / session_name
-
-    def workspace_dir(self) -> Path:
-        return self.session_dir()
-
-    def ensure_session_workspace(self) -> Path:
-        workspace = self.workspace_dir()
-        workspace.mkdir(parents=True, exist_ok=True)
-        self.internal_root_dir().mkdir(parents=True, exist_ok=True)
-        self.session_marker_path().write_text(workspace.name + "\n", encoding="utf-8")
-        return workspace
-
-    def available_sessions(self) -> list[str]:
-        sessions_root = self.sessions_root()
-        if not sessions_root.exists():
-            return []
-        return sorted(path.name for path in sessions_root.iterdir() if path.is_dir())
-
-    def internal_dir(self) -> Path:
-        return self.workspace_dir() / ".konfai_mcp"
-
-    def jobs_dir(self) -> Path:
-        return self.internal_dir() / "jobs"
-
-    def job_dir(self, job_id: str) -> Path:
-        return self.jobs_dir() / job_id
-
-    def job_state_path(self, job_id: str) -> Path:
-        return self.job_dir(job_id) / "job.json"
-
-    def job_manifest_path(self, job_id: str) -> Path:
-        return self.job_dir(job_id) / "manifest.json"
-
-    def job_configs_dir(self, job_id: str) -> Path:
-        return self.job_dir(job_id) / "configs"
-
-    def config_path(self, workflow: str) -> Path:
-        filename = WORKFLOW_CONFIG_FILES.get(workflow)
-        if filename is None:
-            raise ValueError(f"Unsupported workflow: {workflow}")
-        return self.workspace_dir() / filename
-
-    def train_config_path(self) -> Path:
-        return self.config_path("train")
-
-    def prediction_config_path(self) -> Path:
-        return self.config_path("prediction")
-
-    def evaluation_config_path(self) -> Path:
-        return self.config_path("evaluation")
-
-    def statistics_log_path(self) -> Path:
-        return self.workspace_dir() / "Statistics" / "Log.txt"
-
-    def checkpoints_dir(self) -> Path:
-        return self.workspace_dir() / "Checkpoints"
-
-    def predictions_dir(self) -> Path:
-        return self.workspace_dir() / "Predictions"
-
-    def evaluations_dir(self) -> Path:
-        return self.workspace_dir() / "Evaluations"
-
-    def session_workspace_exists(self) -> bool:
-        return self.workspace_dir().exists()
-
-    def ensure_session_workspace_exists(self) -> Path:
-        workspace = self.workspace_dir()
-        if not workspace.exists():
-            raise ValueError(
-                f"Session workspace does not exist for session '{self.current_session}'. Call initialize_session first."
-            )
-        return workspace
-
-    def resolve_workspace_relative_path(self, relative_path: str) -> Path:
-        """Resolve a path inside the session workspace jail.
-
-        Absolute paths are accepted when they resolve inside the workspace, so paths surfaced by
-        job manifests (e.g. config snapshots under ``.konfai_mcp/jobs/``) can be passed back as-is.
-        """
-        # Resolve the workspace too: if the session leaf is a symlink, an unresolved workspace would never
-        # be a parent of the resolved candidate and every in-jail path would be false-rejected.
-        workspace = self.ensure_session_workspace_exists().resolve()
-        candidate = Path(relative_path).expanduser()
-        resolved = candidate.resolve() if candidate.is_absolute() else (workspace / candidate).resolve()
-        if workspace != resolved and workspace not in resolved.parents:
-            raise ValueError("path escapes the session workspace.")
-        return resolved
-
-
-def validate_yaml_content(content: str, filename: str, expected_root: str | None = None) -> dict[str, Any]:
-    """Parse YAML content and optionally enforce one top-level root key."""
-    try:
-        data = YAML_SAFE.load(content)
-    except Exception as exc:  # pragma: no cover - parser exceptions vary
-        raise ValueError(f"{filename} is not valid YAML: {exc}") from exc
-    if not isinstance(data, dict):
-        raise ValueError(f"{filename} must contain a YAML mapping at the top level.")
-    if expected_root is not None and expected_root not in data:
-        raise ValueError(f"{filename} must define the '{expected_root}' root key.")
-    return data
-
-
-def yaml_dump_content(data: dict[str, Any]) -> str:
-    """Serialize a mapping into normalized YAML content."""
-    stream = StringIO()
-    YAML_DUMP.dump(data, stream)
-    return stream.getvalue()
-
-
-def _lint_config_data(data: Any) -> list[dict[str, str]]:
-    """Static lint for silent-failure traps an agent cannot see from the schema alone."""
-    warnings: list[dict[str, str]] = []
-    # Evaluator groups_dest entries bind to GroupTransformMetric, which has no patch_transforms
-    # parameter (and no -1 fill) -- the trap only exists for Trainer/Predictor datasets.
-    if isinstance(data, dict) and isinstance(data.get("Evaluator"), dict):
-        return warnings
-
-    def _walk(node: Any) -> None:
-        if isinstance(node, dict):
-            groups_dest = node.get("groups_dest")
-            if isinstance(groups_dest, dict):
-                for dest_name, dest in groups_dest.items():
-                    if isinstance(dest, dict) and "patch_transforms" not in dest:
-                        warnings.append(
-                            {
-                                "severity": "warning",
-                                "code": "missing_patch_transforms",
-                                "path": f"groups_dest.{dest_name}",
-                                "message": (
-                                    f"groups_dest entry '{dest_name}' omits 'patch_transforms'. Omitting it makes "
-                                    "KonfAI fill this group's tensor with -1 (a segmentation target then crashes "
-                                    "CrossEntropyLoss with 'Target -1 out of bounds'; a regression target is "
-                                    "silently corrupted). Set 'patch_transforms: None' explicitly."
-                                ),
-                            }
-                        )
-            for value in node.values():
-                _walk(value)
-        elif isinstance(node, list):
-            for item in node:
-                _walk(item)
-
-    _walk(data)
-    warnings.extend(_lint_prediction_default_outputs_criterions(data))
-    return warnings
-
-
-_DEFAULT_OUTPUTS_CRITERIONS_GROUP = "Labels"
-
-
-def _contains_key(node: Any, target: str) -> bool:
-    """True when ``target`` appears as a mapping key anywhere in the nested structure."""
-    if isinstance(node, dict):
-        if target in node:
-            return True
-        return any(_contains_key(value, target) for value in node.values())
-    if isinstance(node, list):
-        return any(_contains_key(item, target) for item in node)
-    return False
-
-
-def _predictor_dest_groups(predictor: dict[str, Any]) -> set[str]:
-    """Destination groups the Predictor dataset loads (groups_src.<src>.groups_dest keys)."""
-    groups: set[str] = set()
-    dataset = predictor.get("Dataset")
-    if not isinstance(dataset, dict):
-        return groups
-    groups_src = dataset.get("groups_src")
-    if isinstance(groups_src, dict):
-        for src in groups_src.values():
-            if isinstance(src, dict) and isinstance(src.get("groups_dest"), dict):
-                groups.update(str(name) for name in src["groups_dest"])
-    return groups
-
-
-def _classpath_targets_yaml_model(model: dict[str, Any]) -> bool:
-    """True when Model.classpath points at a .yml/.yaml model builder.
-
-    The YAML model builder defaults ``outputs_criterions`` to None (no injection), so the
-    default-'Labels' binding only happens for Python model classes.
-    """
-    classpath = model.get("classpath")
-    if not isinstance(classpath, str):
-        return False
-    raw = classpath.split("|", maxsplit=1)[-1]
-    return raw.rsplit(".", maxsplit=1)[-1].lower() in {"yml", "yaml"}
-
-
-def _lint_prediction_default_outputs_criterions(data: Any) -> list[dict[str, str]]:
-    """Prediction Model without an explicit outputs_criterions may bind a default referencing 'Labels'.
-
-    When the Predictor Model block omits ``outputs_criterions``, KonfAI's reflection binds the model
-    class's own ``__init__`` default. Builtin models (and plain modules wrapped in ``MinimalModel``)
-    default to a ``TargetCriterionsLoader`` targeting group 'Labels', so their Measure init raises
-    MeasureError when no 'Labels' group is loaded; a custom Network (or a composite like a GAN whose
-    sub-networks own the parameter) may default differently, hence severity=warning, not error.
-    """
-    if not isinstance(data, dict):
-        return []
-    predictor = data.get("Predictor")
-    if not isinstance(predictor, dict):
-        return []
-    model = predictor.get("Model")
-    if not isinstance(model, dict):
-        return []
-    if _classpath_targets_yaml_model(model):
-        return []
-    if _contains_key(model, "outputs_criterions"):
-        return []
-    loaded = _predictor_dest_groups(predictor)
-    if _DEFAULT_OUTPUTS_CRITERIONS_GROUP in loaded:
-        return []
-    loaded_desc = ", ".join(sorted(loaded)) if loaded else "none"
-    return [
-        {
-            "severity": "warning",
-            "code": "prediction_default_outputs_criterions",
-            "path": "Predictor.Model",
-            "message": (
-                "The Predictor Model omits 'outputs_criterions'. For model classes exposing that "
-                "parameter (all builtin models do), KonfAI's reflection binds its default, which "
-                f"references target group '{_DEFAULT_OUTPUTS_CRITERIONS_GROUP}'. This dataset's "
-                f"groups_dest loads [{loaded_desc}] and not '{_DEFAULT_OUTPUTS_CRITERIONS_GROUP}', so "
-                "prediction raises MeasureError when the model initializes its Measure. Set "
-                "'outputs_criterions: {}' on the (sub-)network block that exposes it to disable it "
-                "for pure inference."
-            ),
-        }
-    ]
-
-
-def write_config(path: Path, content: str, overwrite: bool, expected_root: str | None = None) -> dict[str, Any]:
-    """Validate and write one KonfAI YAML config file."""
-    data = validate_yaml_content(content, path.name, expected_root=expected_root)
-    if path.exists() and not overwrite:
-        raise ValueError(f"{path.name} already exists. Set overwrite=True to replace.")
-    path.write_text(content.rstrip() + "\n", encoding="utf-8")
-    result: dict[str, Any] = {
-        "path": str(path),
-        "bytes": path.stat().st_size,
-    }
-    lint = _lint_config_data(data)
-    if lint:
-        result["lint"] = lint
-    return result
-
-
-def read_text(path: Path) -> str:
-    """Return a full text file, or an empty string when it does not exist."""
-    if not path.exists():
-        return ""
-    return path.read_text(encoding="utf-8", errors="replace")
-
-
-def read_text_range(path: Path, max_chars: int = 20000, offset: int = 0) -> dict[str, Any]:
-    """Read a bounded character range of a text file for MCP read-back tools.
-
-    Streams instead of loading the whole file: memory stays proportional to offset+max_chars,
-    so a multi-GB log cannot exhaust the server. ``total_bytes`` is the on-disk size.
-    """
-    if not path.is_file():
-        raise ValueError(f"Not a readable file: {path}")
-    offset = max(offset, 0)
-    max_chars = max(max_chars, 1)
-    with path.open("r", encoding="utf-8", errors="replace") as handle:
-        if offset:
-            handle.read(offset)
-        content = handle.read(max_chars)
-        truncated = bool(handle.read(1))
-    return {
-        "path": str(path),
-        "content": content,
-        "offset": offset,
-        "returned_chars": len(content),
-        "total_bytes": path.stat().st_size,
-        "truncated": truncated,
-    }
-
-
-_BINARY_SNIFF_BYTES = 4096
-_DELIMITED_SUFFIXES = {".csv": ",", ".tsv": "\t"}
-
-
-def read_dataset_sidecar(path: Path, max_lines: int = 200, max_chars: int = 65536) -> dict[str, Any]:
-    """Read a bounded preview of a dataset's non-image text file (CSV/TSV, JSON, YAML, headers, lists).
-
-    Streams at most ``max_chars`` characters (memory never scales with file size) and refuses binary
-    content by sniffing the first bytes for NUL — image volumes and weights belong to
-    ``inspect_dataset``/``preview_volume``, not a text reader. CSV/TSV additionally get a structured
-    preview (header + up to ``max_lines`` rows) so the agent can map label columns to cases without
-    parsing raw text.
-    """
-    if not path.is_file():
-        raise ValueError(f"Not a readable file: {path}")
-    with path.open("rb") as handle:
-        head = handle.read(_BINARY_SNIFF_BYTES)
-    if b"\x00" in head:
-        raise ValueError(
-            f"'{path.name}' looks binary (NUL bytes in the first {_BINARY_SNIFF_BYTES} bytes). "
-            "Use inspect_dataset for volumes/stores and preview_volume for image content."
-        )
-    max_lines = max(max_lines, 1)
-    ranged = read_text_range(path, max_chars=max_chars)
-    lines = ranged["content"].splitlines()
-    line_truncated = len(lines) > max_lines
-    lines = lines[:max_lines]
-    payload: dict[str, Any] = {
-        "ok": True,
-        "path": str(path),
-        "total_bytes": ranged["total_bytes"],
-        "returned_lines": len(lines),
-        "truncated": bool(ranged["truncated"] or line_truncated),
-        "content": "\n".join(lines),
-        "next_actions": ["inspect_dataset", "design_config_strategy"],
-    }
-    delimiter = _DELIMITED_SUFFIXES.get(path.suffix.lower())
-    if delimiter is not None and lines:
-        import csv
-
-        rows = list(csv.reader(StringIO("\n".join(lines)), delimiter=delimiter))
-        if rows:
-            payload["kind"] = "delimited"
-            payload["columns"] = rows[0]
-            payload["rows"] = rows[1:]
-            payload["returned_rows"] = len(rows) - 1
-    return payload
-
-
-def read_text_tail(path: Path, max_lines: int) -> str:
-    """Return the last ``max_lines`` lines of a text file efficiently."""
-    if not path.exists() or max_lines <= 0:
-        return ""
-
-    chunk_size = 8192
-    remaining = max_lines
-    chunks: list[str] = []
-    with path.open("rb") as handle:
-        handle.seek(0, os.SEEK_END)
-        position = handle.tell()
-        while position > 0 and remaining > 0:
-            read_size = min(chunk_size, position)
-            position -= read_size
-            handle.seek(position)
-            chunk = handle.read(read_size).decode("utf-8", errors="replace")
-            chunks.append(chunk)
-            remaining -= chunk.count("\n")
-
-    text = "".join(reversed(chunks))
-    lines = text.splitlines()
-    return "\n".join(lines[-max_lines:])
-
-
-def full_suffix(path: Path) -> str:
-    """Return all suffixes joined together, for example ``.nii.gz``."""
-    return "".join(path.suffixes)
-
-
-def basename_without_suffixes(path: Path) -> str:
-    """Return the basename of ``path`` without compound suffixes."""
-    suffix = full_suffix(path)
-    return path.name[: -len(suffix)] if suffix else path.name
 
 
 def available_templates(examples_root: Path) -> list[str]:
